// src/generated/models/sales-plan-create.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BFF Service
 * Backend For Frontend service aggregating microservices
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: you@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SalesPlanCreate
 */
export interface SalesPlanCreate {
    /**
     * 
     * @type {string}
     * @memberof SalesPlanCreate
     */
    'seller_id': string;
    /**
     * 
     * @type {string}
     * @memberof SalesPlanCreate
     */
    'sales_period': string;
    /**
     * Sales goal amount
     * @type {number}
     * @memberof SalesPlanCreate
     */
    'goal': number;
}